requirements.txt
anyio==3.6.1
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.80.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jeepney==0.8.0
passlib==1.7.4
peewee==3.15.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
